Refactored the front end co-authored-by Lina <woorin.lee1524@gmail.com>
