+++ src/actions/actions.js
@@ -1,4 +1,4 @@
-import * as types from './actionTypes'; 
+import * as types from '../constants/actionTypes'; 
 import axios from 'axios';
