+++ src/pages/LoginPage.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import LoginInformation from '../components/LoginInformation.jsx';
+
+const LoginPage = () => {
+  return (
+    <div>
+      <LoginInformation/>
+    </div>
+  );
+};
+
+export default LoginPage;
