+++ src/pages/SignUpPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import * as actions from '../../actions/actions.js';
+import * as actions from '../actions/actions.js';
 import { bindActionCreators } from 'redux'; 
 import { useNavigate } from 'react-router-dom';
 //import { getState } from 'redux-thunk';
